configs: support psi tracking
Now we have psi tracking support in kernel code, let's make it support
in kconfig.
As of now, we require a boot parameter to enable psi tracking for
performance impact concern. That means we disable it by default, and
user can pass a boot parameter "psi=1" to enable it if they want this
feature.

Signed-off-by: Joseph Qi <joseph.qi@linux.alibaba.com>